backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
sentence-transformers
protobuf==3.20.3
python-docx 
PyMuPDF
python-docx
extract-msg
validators
filetype
scikit-learn
pinecone-client
openai
tiktoken
numpy
